Replace Dummy Code

+++ Models/FavDriver.cs
@@ -18,8 +18,8 @@ namespace UMoveNew.Models
         public string CarNo { get; set; }
 
         public int CarType { get; set; }
-        public double Duration { get; set; }
-        public double Distance { get; set; }
+        public decimal Duration { get; set; }
+        public decimal Distance { get; set; }
 
         public decimal DriverLat { get; set; }
